fix: prepare routes
