+++ src/api/debug.ts
@@ -1,12 +1,27 @@
 import { Router } from 'express'
 import { pino } from '../utils/logger.js'
-import { autoPeeringHandler } from '../auto-peering.cron.js'
+import { getNodesList } from '../utils/utils.js'
+import { helia } from '../helia.js'
 
 const router = Router()
 
-router.get('/debug/autopeering', async (req, res) => {
+router.get('/autopeering', async (req, res) => {
   try {
-    const successPeers = await autoPeeringHandler()
+    const nodes = getNodesList([helia.libp2p.peerId.toString()])
+    pino.logger.info(`Peering nodes list: ${nodes.map((node) => node.name)}`)
+
+    const successPeers: string[] = []
+
+    for await (const node of nodes) {
+      pino.logger.info(`Start peering ${node.name} node (${node.multiAddr})...`)
+      try {
+        await helia.libp2p.dial(node.multiAddr)
+        pino.logger.info(`Successfully peered with ${node.name}`)
+        successPeers.push(node.name)
+      } catch (err) {
+        pino.logger.info(`Peering with ${node.name} failed. Error: ${err.message}`)
+      }
+    }
 
     res.send({
       peeredSuccessfullyTo: successPeers
