+++ src/helia.ts
@@ -9,7 +9,7 @@ import { identify } from '@libp2p/identify'
 import { blockstore, datastore } from './store.js'
 import { getAllowNodesMultiaddrs } from './utils/utils.js'
 import { config } from './config.js'
-import { ConfigNode } from './utils/types.js'
+import { pino } from './utils/logger.js'
 
 export const helia = await createHelia({
   datastore,
@@ -37,6 +37,33 @@ export const helia = await createHelia({
   }
 })
 
-export const ifs = unixfs(helia)
+helia.libp2p.getMultiaddrs().forEach((addr) => {
+  pino.logger.info(`Listening on ${addr.toString()}`)
+})
 
-export const verifiedFetch = await createVerifiedFetch(helia)
+helia.libp2p.addEventListener('peer:discovery', (evt) => {
+  const peer = evt.detail
+  pino.logger.info(`Discovered peer: ${peer.id}`)
+})
+
+helia.libp2p.addEventListener('peer:connect', (evt) => {
+  const peerId = evt.detail
+  pino.logger.info(`Peer connected: ${peerId}`)
+})
+
+helia.libp2p.addEventListener('peer:disconnect', (evt) => {
+  const peerId = evt.detail
+  pino.logger.info(`Peer disconnected: ${peerId}`)
+})
+
+helia.libp2p.addEventListener('start', (event) => {
+  pino.logger.info('Libp2p node started')
+})
+
+helia.libp2p.addEventListener('stop', (event) => {
+  pino.logger.info('Libp2p node stopped')
+})
+
+pino.logger.info(`Helia is running! PeerID: ${helia.libp2p.peerId.toString()}`)
+
+export const ifs = unixfs(helia)
